refactor: extract startup service component boundary
